Fixed bug with debug information
Also, game clients can start server/client through chat (press t for chat).

The debug info bug was caused by the line statements section where we added <preload> as a file entry and lines to instructions on it. That obviously won't work since <preload> isn't a real file. That is why the debugger wasn't working. Now, we don't add debug information for code generated from the preload.

+++ src/BetBat/Util/Utility.cpp
@@ -333,13 +333,13 @@ int PatternMatchFiles(const std::string& pattern, DynamicArray<std::string>* mat
     return num_matched_files;
 }
 std::string TrimLastFile(const std::string& path){
-    int slashI = path.find_last_of("/");
+    size_t slashI = path.find_last_of("/");
     if(slashI==std::string::npos)
-        return "/";
+        return "/"; // just to let you know, code relies on / being returned, linker fails otherwise (i didn't look into why but yeah...)
     return path.substr(0,slashI + 1);
 }
 std::string TrimDir(const std::string& path){
-    int slashI = path.find_last_of("/");
+    size_t slashI = path.find_last_of("/");
     if(slashI==std::string::npos)
         return path;
     return path.substr(slashI+1);
